make sure classlabel is added if not existing
