+++ dataquality/dq_auto/text_classification.py
@@ -10,7 +10,11 @@ from dataquality.dq_auto.base_data_manager import BaseDatasetManager
 from dataquality.dq_auto.tc_trainer import get_trainer
 from dataquality.schemas.split import Split
 from dataquality.schemas.task_type import TaskType
-from dataquality.utils.auto import add_val_data_if_missing, run_name_from_hf_dataset
+from dataquality.utils.auto import (
+    _add_class_label_to_dataset,
+    add_val_data_if_missing,
+    run_name_from_hf_dataset,
+)
 from dataquality.utils.auto_trainer import do_train
 
 a = Analytics(ApiClient, dq.config)
@@ -38,23 +42,6 @@ class TCDatasetManager(BaseDatasetManager):
                 df[c] = df[c].astype("str")
         return df
 
-    def _add_class_label_to_dataset(
-        self, ds: Dataset, labels: Optional[List[str]] = None
-    ) -> Dataset:
-        """Map a not ClassLabel 'label' column to a ClassLabel, if possible"""
-        if "label" not in ds.features or isinstance(ds.features["label"], ClassLabel):
-            return ds
-        labels = labels if labels is not None else sorted(set(ds["label"]))
-        # For string columns, map the label2idx so we can cast to ClassLabel
-        if ds.features["label"].dtype == "string":
-            label_to_idx = dict(zip(labels, range(len(labels))))
-            ds = ds.map(lambda row: {"label": label_to_idx[row["label"]]})
-
-        # https://github.com/python/mypy/issues/6239
-        class_label = ClassLabel(num_classes=len(labels), names=labels)  # type: ignore
-        ds = ds.cast_column("label", class_label)
-        return ds
-
     def _convert_df_to_dataset(
         self, df: pd.DataFrame, labels: Optional[List[str]] = None
     ) -> Dataset:
@@ -69,7 +56,7 @@ class TCDatasetManager(BaseDatasetManager):
         # of the hf DatasetDict will handle this missing label column if it's an
         # issue. See `_validate_dataset_dict`
         ds = Dataset.from_pandas(df_copy)
-        return self._add_class_label_to_dataset(ds, labels)
+        return _add_class_label_to_dataset(ds, labels)
 
     def _validate_dataset_dict(
         self,
@@ -96,7 +83,7 @@ class TCDatasetManager(BaseDatasetManager):
             if key not in inference_names:
                 assert "label" in ds.features, "Dataset must have column `label`"
                 if not isinstance(ds.features["label"], ClassLabel):
-                    ds = self._add_class_label_to_dataset(ds, labels)
+                    ds = _add_class_label_to_dataset(ds, labels)
             if "id" not in ds.features:
                 ds = ds.add_column("id", list(range(ds.num_rows)))
             clean_dd[key] = ds
