+++ dataquality/utils/auto.py
@@ -188,3 +188,21 @@ def set_global_logging_level(
     for name in logging.root.manager.loggerDict:
         if re.match(prefix_re, name):
             logging.getLogger(name).setLevel(level)
+
+
+def _add_class_label_to_dataset(
+    ds: Dataset, labels: Optional[List[str]] = None
+) -> Dataset:
+    """Map a not ClassLabel 'label' column to a ClassLabel, if possible"""
+    if "label" not in ds.features or isinstance(ds.features["label"], ClassLabel):
+        return ds
+    labels = labels if labels is not None else sorted(set(ds["label"]))
+    # For string columns, map the label2idx so we can cast to ClassLabel
+    if ds.features["label"].dtype == "string":
+        label_to_idx = dict(zip(labels, range(len(labels))))
+        ds = ds.map(lambda row: {"label": label_to_idx[row["label"]]})
+
+    # https://github.com/python/mypy/issues/6239
+    class_label = ClassLabel(num_classes=len(labels), names=labels)  # type: ignore
+    ds = ds.cast_column("label", class_label)
+    return ds
